Fixed normalization, cleaned up.
The big deal is using the CMD space defined by the data RGB
normalization to select the number of sim RGB stars. This fixed the
normalization and the LFs match.

I raises a new issue of needing to properly nail down the spread code.
Before, was only using stars in the cmd space that were RGB. The offset
in the LF was from NON-RGB stars spread to that region…

Also:
got rid of import * magic
got rid of some unused modules