Major changes!
And there will be more, the splitting of AGBTrackUtils and fileIO is a
bit stilly.

All kwargs from the input file are gone and replaced by input_file
class which holds them and sets their defaults.

Finding quessient phases and assigning slopes is more robust.

All calculations are now done in AGBTracks class, not scirpted in
do_everything.

The speed is faster, but still sucks. genfromtxt calls are out of hand,
as are ~15*60 matplotlib saves… not sure how to speed it up besides
multiprocessing.